requirements.txt
firebase-admin==5.0.1
gunicorn 
python-dotenv == 0.19.1
Flask==2.0.1
python-dotenv == 0.19.1
requests